backend/requirements.txt
strawberry-graphql==0.216.1
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.4.2
python-multipart==0.0.6